taint/test/mochitest/test_structured_clone.html

<!DOCTYPE HTML>
<html>
  <head>
    <meta charset="utf-8">
    <title>Check Location Sources</title>
    <script src="/tests/SimpleTest/SimpleTest.js"></script>
    <link rel="stylesheet" href="/tests/SimpleTest/test.css"/>
    <script>

      function check_tainted(str) {
          ok(str.taint.length > 0, "Check tainted");
      }

      function check_untainted(str) {
          ok(str.taint.length == 0, "Check untainted");
      }

      add_task(async function structuredCloneDirect() {
        let s = String.tainted("test");
        let sc = window.structuredClone(s);
        check_tainted(sc);
      });

      add_task(async function structuredCloneMixed() {
        let s = "this is a " + String.tainted("test");
        let sc = window.structuredClone(s);
        check_tainted(sc);
      });

      add_task(async function structuredCloneObject() {
        let s = "this is a " + String.tainted("test");
        let o = { k: s, v: 134 };
        let oc = window.structuredClone(o);
        check_tainted(o.k);
      });

    </script>
  </head>
  <body>
    <p id="display"></p>
    <div id="content" style="display: none"></div>
    <pre id="test"></pre>
  </body>
</html>
